+++ soya/src/data/redux/segment/map/MapSegment.js
@@ -1,6 +1,6 @@
 import Segment from '../../Segment.js';
 import QueryOptionUtil from '../QueryOptionUtil.js';
-import MapActionCreator from './MapActionCreator.js';
+import ActionNameUtil from '../ActionNameUtil.js';
 
 /**
  * Promise implementation, in local variable so that our code can be natural.
@@ -23,9 +23,24 @@ export default class MapSegment extends Segment {
   _queries;
 
   /**
-   * @type {MapActionCreator}
+   * @type {Object}
    */
   _actionCreator;
+
+  /**
+   * @type {string}
+   */
+  _loadActionType;
+
+  /**
+   * @type {string}
+   */
+  _initActionType;
+
+  /**
+   * @type {string}
+   */
+  _cleanActionType;
 
   /**
    * @param config
@@ -35,32 +50,55 @@ export default class MapSegment extends Segment {
   constructor(config, provider, PromiseImpl) {
     super(config, provider, PromiseImpl);
     Promise = PromiseImpl;
-    this._actionCreator = this._createActionCreator();
-    if (!(this._actionCreator instanceof MapActionCreator)) {
-      throw new Error('MapSegment must be used in tandem with MapActionCreator.');
-    }
+
+    // Since segment name is guaranteed never to clash by ReduxStore, we can
+    // safely use segment name as action type.
+    var id = this.constructor.id();
+    this._loadActionType = ActionNameUtil.generate(id, 'LOAD');
+    this._initActionType = ActionNameUtil.generate(id, 'INIT');
+    this._cleanActionType = ActionNameUtil.generate(id, 'CLEAN');
+
     this._queries = {};
+    this._promises = {};
+    this._actionCreator = {
+      load: (query) => {
+        var queryId = this._generateQueryId(query);
+        return this._createThunk(query, queryId);
+      }
+    };
   }
 
   /**
-   * Creates MapActionCreator implementation. Will only be called once.
+   * Generates a unique string representing the given query. Same query must
+   * generate identical strings. Query ID is used by ReduxStore and Segment
+   * to recognize identical queries.
    *
-   * ABSTRACT: To be overridden by child implementation.
+   * ABSTRACT: To be overridden by child implementations.
    *
-   * @return {MapActionCreator}
+   * @param {any} query
+   * @return {string}
+   * @private
    */
-  _createActionCreator() {
-
+  _generateQueryId(query) {
+    throw new Error('User must override this _generateQueryId method! Instance: ' + this + '.');
   }
 
   /**
-   * Child classes may want override merging behavior.
+   * Fetches query result from external source, returns a Redux thunk function
+   * that may be wrapped with QueryDependencies.
    *
-   * @param {Object} optionA
-   * @param {Object} optionB
+   * IMPORTANT NOTE: The thunk function must do the dispatch, and return a
+   * Promise that resolves *after* dispatch is done.
+   *
+   * ABSTRACT: To be overridden by child implementations.
+   *
+   * @param {any} query
+   * @param {string} queryId
+   * @return {Function}
+   * @private
    */
-  _mergeOptions(optionA, optionB) {
-    return QueryOptionUtil.mergeOptions(optionA, optionB);
+  _createThunk(query, queryId) {
+    throw new Error('User must override _fetch method! Instance: ' + this + '.');
   }
 
   /**
@@ -68,34 +106,46 @@ export default class MapSegment extends Segment {
    * already registered, we merge its options.
    *
    * @param {any} query
-   * @param {Object} options
    * @return {string}
    */
-  _registerQuery(query, options) {
-    options = QueryOptionUtil.initOptions(options);
-    var queryId = this._actionCreator._generateQueryId(query);
-    if (this._queries.hasOwnProperty(queryId)) {
-      this._queries[queryId].options = this._mergeOptions(
-        options, this._queries[queryId].options)
-    } else {
-      this._queries[queryId] = {
-        query: query,
-        options: options
-      };
+  _registerQuery(query) {
+    var queryId = this._generateQueryId(query);
+    if (!this._queries.hasOwnProperty(queryId)) {
+      this._queries[queryId] = query;
     }
     return queryId;
   }
 
   /**
    * @param {string} queryId
-   * @return {Object | Function}
+   * @return {Object | Thunk}
    */
-  _createHydrateAction(queryId) {
-    return this._actionCreator.createLoadAction(
-      this._queries[queryId].query,
-      this._queries[queryId].options,
-      true
-    );
+  _createLoadAction(queryId) {
+    var query = this._queries[queryId];
+    return this._createThunk(query, queryId);
+  }
+
+  /**
+   * Creates an action object with the given state payload.
+   *
+   * IMPORTANT NOTE: Please make sure that you return a *new* object, as redux
+   * store states are supposed to be immutable.
+   *
+   * @param {string} queryId
+   * @param {void | any} payload
+   * @param {void | Array<any>} errors
+   * @return {Object}
+   */
+  _createLoadActionObject(queryId, payload, errors) {
+    return {
+      type: this._loadActionType,
+      queryId: queryId,
+      payload: {
+        data: payload,
+        errors: errors,
+        loaded: true
+      }
+    };
   }
 
   /**
@@ -103,18 +153,28 @@ export default class MapSegment extends Segment {
    * @return {Object}
    */
   _createInitAction(queryId) {
-    return this._actionCreator._createInitActionObject(queryId);
+    return {
+      type: this._initActionType,
+      queryId: queryId,
+      payload: {
+        data: null,
+        errors: null,
+        loaded: false
+      }
+    };
   }
 
   /**
    * @return {Object}
    */
   _createCleanAction() {
-    return this._actionCreator._createCleanAction();
+    return {
+      type: this._cleanActionType
+    };
   }
 
   /**
-   * @return {MapActionCreator}
+   * @return {Object}
    */
   _getActionCreator() {
     return this._actionCreator;
@@ -165,9 +225,9 @@ export default class MapSegment extends Segment {
    * @return {Function}
    */
   _getReducer() {
-    var loadActionType = this._actionCreator._getLoadActionType();
-    var initActionType = this._actionCreator._getInitActionType();
-    var cleanActionType = this._actionCreator._getCleanActionType();
+    var loadActionType = this._loadActionType;
+    var initActionType = this._initActionType;
+    var cleanActionType = this._cleanActionType;
     return (state, action) => {
       // If state is undefined, return initial state.
       if (!state) state = {};