+++ test/integration/src/segments/user/UserSegment.js
@@ -1,12 +1,36 @@
 import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment.js';
-import UserActionCreator from './UserActionCreator.js';
+import Thunk from 'soya/lib/data/redux/Thunk.js';
+
+// TODO: Figure out how to do polyfill.
+// TODO: Figure out how to load client-side libraries like jQuery!
+import { Promise } from 'es6-promise';
+import request from 'superagent';
 
 export default class UserSegment extends MapSegment {
   static id() {
     return 'user';
   }
 
-  _createActionCreator() {
-    return new UserActionCreator(UserSegment.id());
+  _generateQueryId(query) {
+    return query.username;
+  }
+
+  _createThunk(query, queryId) {
+    var thunk = new Thunk(queryId);
+    thunk.func = (dispatch) => {
+      var result = new Promise((resolve, reject) => {
+        request.get('http://localhost:8000/api/user/' + query.username).end((err, res) => {
+          if (res.ok) {
+            var payload = JSON.parse(res.text);
+            dispatch(this._createLoadActionObject(queryId, payload));
+            resolve();
+          } else {
+            reject(new Error('Unable to fetch user data!'));
+          }
+        });
+      });
+      return result;
+    };
+    return thunk;
   }
 }